@@ -63,7 +63,7 @@ dependencies:
   table_calendar: ^3.2.0
   sliding_up_panel: ^2.0.0+1
   share_plus: ^10.1.4 
-  webview_flutter: ^4.13.0
+  youtube_player_flutter: ^9.1.3
  
 
 dev_dependencies:
